+++ src/frame_compare/utils/__init__.py
@@ -2,6 +2,12 @@
 
 from frame_compare.utils.logging import configure_logging, get_run_id, new_run_id
 from frame_compare.utils.perf import is_perf_enabled, perf_span
+from frame_compare.utils.progress import (
+    LogProgressReporter,
+    NullProgressReporter,
+    ProgressReporter,
+    RichProgressReporter,
+)
 
 __all__ = [
     "configure_logging",
@@ -9,4 +15,8 @@ __all__ = [
     "new_run_id",
     "is_perf_enabled",
     "perf_span",
+    "ProgressReporter",
+    "NullProgressReporter",
+    "RichProgressReporter",
+    "LogProgressReporter",
 ]